ag grid in lazy module
