+++ testing/module-size-angular/src/app/app.routes.ts
@@ -1,3 +1,3 @@
-import { Routes } from '@angular/router';
+import type { Routes } from '@angular/router';
 
 export const routes: Routes = [];
